RESKC-1288: Fix state validation when DC is used as the state. (#91)
When adding an organization with a Washington DC address to an S2S proposal, KC flags the state field as an error. DC is an active entry in KC's State table, and it also exists in the XML schema here that's used by the PerformanceSite form here:
http://apply07.grants.gov/apply/system/schemas/UniversalCodes-V2.0.xsd
So I'm not sure why KC is flagging this as an error. Changing the state to a "standard" one like FL resolves the error.
Reproduced this issue on res-demo2 with proposal 399; screenshots are attached.
Also in Demo 1 Proposal 1322
Errors re state occur identified so far:
SF 424 R&R 2.0
Senior/Key Person Expanded 2.0