RESKC-1312: not all extra key persons are included on the period when the PI is not listed on the budget period.
To reproduce:

Create a proposal and add some PI and some Key Personnel to the proposal.
Add an S2S opportunity with the RR Budget 1-3 form. (PA-DD-000)
Add a budget justification attachment
Create a budget
Add enough additional Budget Persons to exceed the 8 row detailed in the RR Budget form.
Assign Personnel:
DO NOT Add the PI to the budget.
Add all but the PI and one other Key person to Period 1, making sure they are all in the Senior Personnel budget category.
Generate periods
Add the remaining Key Person (NOT THE PI) to P2 and assign to later periods.
From Budget Versions: Print the Salary Report to use for comparison
Return to proposal: S2S – Print the RR budget form.
Compare the Salary Report to the Budget and Additional Budget Persons print out.
Notice that the ‘9th’ person on the salary report for each period is missing from the Senior person section of the Budget form, and is not included in the Additional Budget Person’s sheet.